+++ app/src/main/java/projects/medicationtracker/Fragments/MyMedicationsFragment.java
@@ -1,8 +1,6 @@
 package projects.medicationtracker.Fragments;
 
-import static projects.medicationtracker.AddMedication.MINUTES_IN_DAY;
 import static projects.medicationtracker.Helpers.DBHelper.DATE_FORMAT;
-import static projects.medicationtracker.Helpers.DBHelper.TIME_FORMAT;
 import static projects.medicationtracker.MainActivity.preferences;
 
 import android.content.Intent;
@@ -15,7 +13,6 @@ import android.widget.TextView;
 
 import androidx.fragment.app.Fragment;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
@@ -23,7 +20,6 @@ import java.util.Locale;
 
 import projects.medicationtracker.AddMedication;
 import projects.medicationtracker.Helpers.DBHelper;
-import projects.medicationtracker.Helpers.TimeFormatting;
 import projects.medicationtracker.MedicationNotes;
 import projects.medicationtracker.R;
 import projects.medicationtracker.SimpleClasses.Medication;
@@ -92,8 +88,8 @@ public class MyMedicationsFragment extends Fragment {
 
         medication.setTimes(dateTimes);
 
-        if (medication.getDosage() == (int) medication.getDosage()) {
-            dosageVal = String.format(Locale.getDefault(), "%d", (int) medication.getDosage());
+        if (medication.getDosage() == medication.getDosage()) {
+            dosageVal = String.format(Locale.getDefault(), "%d", medication.getDosage());
         } else {
             dosageVal = String.valueOf(medication.getDosage());
         }
